requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
frozenlist==1.4.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
gspread==6.1.0
httplib2==0.22.0
humanfriendly==10.0
idna==3.7
Mako==1.3.3
MarkupSafe==2.1.5
multidict==6.0.5
oauth2client==4.1.3
oauthlib==3.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pyTelegramBotAPI==4.17.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
StrEnum==0.4.15
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
yarl==1.9.4
